20 second run with lego motors
