+++ lib/control.h
@@ -1,43 +1,16 @@
-  void stop_box(int mode)
+//LF
+  void stop_sensor(int sn, int mode) // 0 - stop when high, 1 - stop when low
   {
     async_reset();
+    int logic = 1;
 
-    while(digitalRead(BOX_LIMIT_PIN) == mode)
+    while(logic)
     {
-      async();
-    }
-  }
+      read_sv();
 
-  void stop_time(int mils)
-  {
-    async_reset();
-    int dt = 0;
+      if(mode == 0){logic = (svals[sn] < 50);}
+      else if(mode == 1){logic = (svals[sn] > 50);}    
 
-    while(dt < mils)
-    {
-      dt += em;
-      async();
-    }
-  }
-
-  void stop_corner()
-  {
-    async_reset();
-    read_sv();
-
-    while( density < 4 || (svals[0] < 50 && svals[NUMLSENSORS - 1] < 50) )
-    {
-      async();
-    }
-  }
-  void stop_deg(float a)
-  {
-    async_reset();
-    motion[0] = 0;
-    int dir = sign_f(a);
-
-    while(motion[0]*dir < abs(a))
-    {
       async();
     }
   }
@@ -49,6 +22,17 @@
     while(density != 0)
     {
       read_sv();
+      async();
+    }
+  } 
+
+//DISTANCE
+  void stop_box(int mode)
+  {
+    async_reset();
+
+    while(digitalRead(BOX_LIMIT_PIN) == mode)
+    {
       async();
     }
   }
@@ -64,18 +48,78 @@
     while(dd*dir < abs(dist)){async();}
   }
 
-  void stop_sensor(int sn, int mode) // 0 - stop when high, 1 - stop when low
+//INTERNAL
+  void stop_time(int mils)
   {
     async_reset();
-    int logic = 1;
+    int dt = 0;
 
-    while(logic)
+    while(dt < mils)
     {
-      read_sv();
-
-      if(mode == 0){logic = (svals[sn] < 50);}
-      else if(mode == 1){logic = (svals[sn] > 50);}    
-
+      dt += em;
       async();
     }
+  }
+
+  void stop_deg(float a)
+  {
+    async_reset();
+    motion[0] = 0;
+    int dir = sign_f(a);
+
+    while(motion[0]*dir < abs(a))
+    {
+      async();
+    }
+  }
+
+//EXTENDED
+  void stop_corner()
+  {
+    async_reset();
+    read_sv();
+
+    while( density < 4 || (svals[0] < 50 && svals[NUMLSENSORS - 1] < 50) )
+    {
+      async();
+    }
+  }
+
+  void no_state()
+  {
+    async_state = -1;
+  }
+
+  void turn_r()
+  {
+    no_state();
+    rotate(200, 0);
+    read_sv();
+
+    async_reset();
+
+    while(!eval_line(0x80, 0x80))
+    {
+      async();
+    }
+
+    lf_settle(200);
+    stop_time(TIME2SETTLE);
+  }
+
+  void turn_l()
+  {
+    no_state();
+    rotate(-200, 0);
+    read_sv();
+
+    async_reset();
+
+    while(!eval_line(0x01, 0x01))
+    {
+      async();
+    }
+
+    lf_settle(200);
+    stop_time(TIME2SETTLE);
   }
