+++ lib/toolbox.h
@@ -1,0 +1,21 @@
+  int sign_f(float val)
+  {
+    if(val < 0){return -1;}
+    else if(val > 0){return 1;}
+    else{return 0;}
+  }
+
+  int sign_i(int val)
+  {
+    if(val < 0){return -1;}
+    else if(val > 0){return 1;}
+    else{return 0;}
+  }
+
+  void blink_led()
+  {
+    digitalWrite(13, HIGH);
+    delay(300);
+    digitalWrite(13, LOW);
+    delay(300);
+  }
